+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# [next]
+- Adds `onContactExit` in `Sensor` mixin.
+
 # [2.8.0]
 - Update flame to `1.3.0`